+++ instrumentation/opentelemetry-instrumentation-asgi/src/opentelemetry/instrumentation/asgi/__init__.py
@@ -218,7 +218,9 @@ def set_status_code(span, status_code):
         )
     else:
         span.set_attribute(SpanAttributes.HTTP_STATUS_CODE, status_code)
-        span.set_status(Status(http_status_to_status_code(status_code)))
+        span.set_status(
+            Status(http_status_to_status_code(status_code, server_span=True))
+        )
 
 
 def get_default_span_details(scope: dict) -> Tuple[str, dict]: