+++ instrumentation/opentelemetry-instrumentation-tornado/src/opentelemetry/instrumentation/tornado/__init__.py
@@ -345,7 +345,9 @@ def _finish_span(tracer, handler, error=None):
 
     if ctx.span.is_recording():
         ctx.span.set_attribute(SpanAttributes.HTTP_STATUS_CODE, status_code)
-        otel_status_code = http_status_to_status_code(status_code)
+        otel_status_code = http_status_to_status_code(
+            status_code, server_span=True
+        )
         otel_status_description = None
         if otel_status_code is StatusCode.ERROR:
             otel_status_description = reason